MongoDB for Pycharm
We have to store all our data in MongoDB. Fortunately, this database
accepts "documents" which are "JSON objects" as input. So we don't have
to process our data a lot for storing it.
The first code snippet is an example I just tried to learn how it work.
I let it be there in case you wanna see it.
MongoDB can be easily installed and configured. Make sure you go through
all steps of installation and configuration noted in the official site
^_^